* paedml50-upgrade:    - --force_Parameter hinzugefügt, um Versionscheck zu überspringen.  * paedml50-upgrade.sh:    - viele Verbesserungen bzgl. des Upgrades von Testingsystemen.    - Skript kann jetzt nach einem abgebrochenen Upgrade noch einmal aufgerufen werden, um wieder in einen konsistenten Zustand zu gelangen.

+++ share/upgrade/paedml50-upgrade.sh
@@ -17,13 +17,19 @@ QUOTDYNTPLDIR=$DYNTPLDIR/18_quota
 HORDDYNTPLDIR=$DYNTPLDIR/21_horde3
 NAGIDYNTPLDIR=$DYNTPLDIR/22_nagios
 FREEDYNTPLDIR=$DYNTPLDIR/55_freeradius
-dpkg -s samba | grep ^Version | grep -q 3.0.24 || SAMBAPKGs="samba samba-common"
+OPENML=`dpkg -l | grep "schulkonsole-templates-openlml" | grep ^i`
 PYKOTA=`dpkg -l | grep "linuxmuster-pk " | grep ^i`
-[ -n "$PYKOTA" ] && PYKOTAPKGS="linuxmuster-pykota linuxmuster-pk linuxmuster-pkpgcounter python-egenix-mxtools python-egenix-mxdatetime"
+BITTORRENT=`dpkg -l | grep " bittorrent " | grep ^i`
 FREERADIUS=`dpkg -l | grep linuxmuster-freeradius | grep ^i`
+PHPMYADMIN=`dpkg -l | grep phpmyadmin | grep ^i`
+PHPPGADMIN=`dpkg -l | grep phppgadmin | grep ^i`
 COPSPOT=`dpkg -l | grep linuxmuster-ipcop-addon-copspot | grep ^i`
-SOPHOPKGS=`dpkg -l | grep sophomorix | grep ^i | awk '{ print $2 }'`
-PKGSTOREMOVE="linuxmuster-linbo linuxmuster-freeradius linux-image-server nagios2 linuxmuster-nagios-base mindi mondo postgresql-7.4 postgresql-8.1 $SOPHOPKGS $PYKOTAPKGS $SAMBAPKGS"
+PKGSTOREMOVE="linuxmuster-freeradius linux-image-server phpmyadmin phppgadmin \
+              linuxmuster-schulkonsole-templates-openlml mindi mondo nagios2 \
+              linuxmuster-nagios-base postgresql-7.4 postgresql-8.1 samba \
+              linuxmuster-pkpgcounter python-egenix-mxtools python-egenix-mxdatetime \
+              linuxmuster-pykota linuxmuster-pk samba-common samba-common-bin \
+              sophomorix2 sophomorix-base sophomorix-pgldap"
 PKGREPOS="ftp.de.debian.org/debian/ \
           ftp.de.debian.org/debian-volatile/ \
           security.debian.org \
@@ -79,7 +85,7 @@ for i in servername domainname internmask internsubrange imaging sambasid workgr
     echo "imaging=rembo" >> $NETWORKSETTINGS
    fi
   else
-   echo "    Fatal! $i ist nicht gesetzt!"
+   echo "    Fehler! $i ist nicht gesetzt!"
    exit 1
   fi
  fi
@@ -92,13 +98,13 @@ internbc=`echo $internsubrange | cut -f2 -d"-"`
 serverip=10.$internsub.1.1
 echo "  * serverip=$serverip"
 if ! validip "$serverip"; then
- echo "    Fatal! serverip ist ungueltig!"
+ echo "    Fehler! serverip ist ungueltig!"
  exit 1
 fi
 ipcopip=10.$internsub.1.254
 echo "  * ipcopip=$ipcopip"
 if ! validip "$ipcopip"; then
- echo "    Fatal! ipcopip ist ungueltig!"
+ echo "    Fehler! ipcopip ist ungueltig!"
  exit 1
 fi
 broadcast=10.$internbc.255.255
@@ -117,11 +123,16 @@ for i in `psql -t -l -U postgres | awk '{ print $1 }'`; do
  case $i in
   postgres|template0|template1) continue ;;
  esac
+ if [ -s "$BACKUPDIR/$i/$i.lenny-upgrade.pgsql.gz" ]; then
+  echo "Sicherung von $i-Datenbank gefunden. Lasse $i aus!"
+  continue
+ fi
  echo "Sichere $i-Datenbank nach $BACKUPDIR/$i/$i.lenny-upgrade.pgsql.gz ..."
  RC=1
  pg_dump --encoding=UTF8 -U postgres $i > /var/tmp/$i.lenny-upgrade.pgsql ; RC="$?"
  if [ "$RC" != "0" ]; then
   echo "Fehler bei der Sicherung der $i-Datenbank!"
+  rm -f /var/tmp/$i.lenny-upgrade.pgsql
   exit 1
  fi
  mkdir -p "$BACKUPDIR/$i"
@@ -129,6 +140,7 @@ for i in `psql -t -l -U postgres | awk '{ print $1 }'`; do
  gzip -c9 /var/tmp/$i.lenny-upgrade.pgsql > $BACKUPDIR/$i/$i.lenny-upgrade.pgsql.gz ; RC="$?"
  if [ "$RC" != "0" ]; then
   echo "Fehler bei der Komprimierung der $i-Datenbank!"
+  rm -f $BACKUPDIR/$i/$i.lenny-upgrade.pgsql.gz
   exit 1
  fi
 done
@@ -138,32 +150,35 @@ echo
 echo "#####################"
 echo "# apt-Konfiguration #"
 echo "#####################"
-cp /etc/apt/sources.list /etc/apt/sources.list.lenny-upgrade
+[ -e /etc/apt/sources.list.lenny-upgrade ] || cp /etc/apt/sources.list /etc/apt/sources.list.lenny-upgrade
 [ -e /etc/apt/apt.conf ] && mv /etc/apt/apt.conf /etc/apt/apt.conf.lenny-upgrade
-mv /etc/apt/sources.list.d /etc/apt/sources.list.d.lenny-upgrade
+[ -d /etc/apt/sources.list.d.lenny-upgrade ] || mv /etc/apt/sources.list.d /etc/apt/sources.list.d.lenny-upgrade
 cp -a $STATICTPLDIR/etc/apt/* /etc/apt
 
+tweak_apt() {
+ export DEBIAN_FRONTEND=noninteractive
+ export DEBIAN_PRIORITY=critical
+ export DEBCONF_TERSE=yes
+ export DEBCONF_NOWARNINGS=yes
+ echo 'DPkg::Options {"--force-configure-any";"--force-confmiss";"--force-confold";"--force-confdef";"--force-bad-verify";"--force-overwrite";};' > /etc/apt/apt.conf.d/99upgrade
+ echo 'APT::Get::AllowUnauthenticated "true";' >> /etc/apt/apt.conf.d/99upgrade
+}
+
 # force apt to do an unattended upgrade
-export DEBIAN_FRONTEND=noninteractive
-export DEBIAN_PRIORITY=critical
-export DEBCONF_TERSE=yes
-export DEBCONF_NOWARNINGS=yes
-echo 'DPkg::Options {"--force-confold";"--force-confdef";"--force-bad-verify";"--force-overwrite";};' > /etc/apt/apt.conf.d/99upgrade
-echo 'APT::Get::AllowUnauthenticated "true";' >> /etc/apt/apt.conf.d/99upgrade
 echo "Aktualisiere Paketlisten ..."
-if aptitude update; then
- echo "Ok."
+tweak_apt
+if ! aptitude update; then
  echo
-else
- echo "Kann Paketlisten nicht aktualisieren."
+ echo "Fehler: Kann Paketlisten nicht aktualisieren."
  exit 1
 fi
+echo
 
 
 echo "#########################"
 echo "# Pakete deinstallieren #"
 echo "#########################"
-echo -e "\n\n" | aptitude -y remove $PKGSTOREMOVE
+aptitude -y remove $PKGSTOREMOVE
 echo
 
 
@@ -173,19 +188,19 @@ echo "###############################"
 # uml utilities
 echo " uml-utilities ..."
 CONF=/etc/default/uml-utilities
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # bootlogd
 echo " bootlogd ..."
 CONF=/etc/default/bootlogd
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # slapd
 echo " slapd ..."
 for i in /etc/ldap/slapd.conf /etc/default/slapd /var/lib/ldap/DB_CONFIG; do
- cp $i $i.lenny-upgrade
+ [ -e "$i.lenny-upgrade" ] || cp $i $i.lenny-upgrade
  if stringinstring slapd.conf $i; then
   ldapadminpw=`grep ^rootpw $i | awk '{ print $2 }'`
   sed -e "s/@@message1@@/${message1}/
@@ -206,12 +221,12 @@ chmod 600 /var/lib/ldap/*
 # smbldap-tools
 echo " smbldap-tools ..."
 CONF=/etc/smbldap-tools/smbldap.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e "s/@@sambasid@@/${sambasid}/
 	       s/@@workgroup@@/${workgroup}/
 	       s/@@basedn@@/${basedn}/" $LDAPDYNTPLDIR/`basename $CONF` > $CONF
 CONF=/etc/smbldap-tools/smbldap_bind.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e "s/@@message1@@/${message1}/
 	       s/@@message2@@/${message2}/
 	       s/@@message3@@/${message3}/
@@ -222,30 +237,32 @@ chmod 600 ${CONF}*
 # apache2
 echo " apache2 ..."
 CONF=/etc/apache2/apache2.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 rm -f /etc/apache2/mods-enabled/mime_magic.*
 
 # nagios2
-CONF=/etc/nagios2/resource.cfg
-[ -e "$CONF" ] || touch $CONF
-CONF=/etc/apache2/conf.d/nagios2.conf
-if [ -e "$CONF" ]; then
- backup_file $CONF
- rm -f $CONF
+if [ -d /etc/nagios2 ]; then
+ echo " nagios2 ..."
+ CONF=/etc/nagios2/resource.cfg
+ [ -e "$CONF" ] || touch $CONF
+ CONF=/etc/apache2/conf.d/nagios2.conf
+ if [ -e "$CONF" ]; then
+  backup_file $CONF
+  rm -f $CONF
+ fi
 fi
-[ -d /etc/nagios2/conf.d ] && cp -a /etc/nagios2/conf.d /etc/nagios2/conf.d.lenny-upgrade
 
 # saslauthd
 echo " saslauthd ..."
 CONF=/etc/default/saslauthd
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # dhcp
 echo " dhcp ..."
 CONF=/etc/dhcp3/dhcpd.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e "s/@@servername@@/${servername}/g
         s/@@domainname@@/${domainname}/g
         s/@@serverip@@/${serverip}/g
@@ -259,7 +276,7 @@ sed -e "s/@@servername@@/${servername}/g
 echo " bind9 ..."
 for i in db.10 db.linuxmuster named.conf.linuxmuster; do
  CONF=/etc/bind/$i
- cp $CONF $CONF.lenny-upgrade
+ [ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
  sed -e "s/@@servername@@/${servername}/g
          s/@@domainname@@/${domainname}/g
          s/@@serverip@@/${serverip}/g
@@ -271,10 +288,10 @@ rm -f /etc/bind/*.jnl
 # horde 3
 echo " horde3 ..."
 CONF=/etc/horde/horde3/registry.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 CONF=/etc/horde/horde3/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 hordepw="$(grep "^\$conf\['sql'\]\['password'\]" $CONF | awk -F\' '{ print $6 }')"
 sed -e "s/\$conf\['auth'\]\['admins'\] =.*/\$conf\['auth'\]\['admins'\] = array\('$WWWADMIN'\);/
         s/\$conf\['problems'\]\['email'\] =.*/\$conf\['problems'\]\['email'\] = '$WWWADMIN@$domainname';/
@@ -283,35 +300,35 @@ sed -e "s/\$conf\['auth'\]\['admins'\] =.*/\$conf\['auth'\]\['admins'\] = array\
         s/\$conf\['sql'\]\['password'\] =.*/\$conf\['sql'\]\['password'\] = '$hordepw';/" $STATICTPLDIR/$CONF > $CONF
 # imp
 CONF=/etc/horde/imp4/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 CONF=/etc/horde/imp4/servers.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cyradmpw="$(cat /etc/imap.secret)"
 sed -e "s/'@@servername@@.@@domainname@@'/'$servername.$domainname'/g
         s/'@@domainname@@'/'$domainname'/g
         s/'@@cyradmpw@@'/'$cyradmpw'/" $HORDDYNTPLDIR/imp4.`basename $CONF` > $CONF
 # ingo
 CONF=/etc/horde/ingo1/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 # kronolith
 CONF=/etc/horde/kronolith2/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e "s/\$conf\['storage'\]\['default_domain'\] =.*/\$conf\['storage'\]\['default_domain'\] = '$domainname';/
         s/\$conf\['reminder'\]\['server_name'\] =.*/\$conf\['reminder'\]\['server_name'\] = '$servername.$domainname';/
         s/\$conf\['reminder'\]\['from_addr'\] =.*/\$conf\['reminder'\]\['from_addr'\] = '$WWWADMIN@$domainname';/" $STATICTPLDIR/$CONF > $CONF
 # mnemo
 CONF=/etc/horde/mnemo2/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 # nag
 CONF=/etc/horde/nag2/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 # turba
 CONF=/etc/horde/turba2/conf.php
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 # permissions
 chown root:www-data /etc/horde -R
@@ -320,38 +337,38 @@ find /etc/horde -type f -exec chmod 440 '{}' \;
 # php5
 echo " php5 ..."
 CONF=/etc/php5/conf.d/paedml.ini
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 CONF=/etc/php5/apache2/php.ini
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 CONF=/etc/php5/cli/php.ini
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # mindi
 echo " mindi ..."
 CONF=/etc/mindi/mindi.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # warnquota
 echo " warnquota ..."
 CONF=/etc/warnquota.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e "s|@@administrator@@|$ADMINISTRATOR|g
         s|@@domainname@@|$domainname|g" $QUOTDYNTPLDIR/$(basename $CONF) > $CONF
 
 # webmin
 echo " webmin ..."
 CONF=/etc/webmin/config
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 cp $STATICTPLDIR/$CONF $CONF
 
 # fixing backup.conf
 echo " backup ..."
 CONF=/etc/linuxmuster/backup.conf
-cp $CONF $CONF.lenny-upgrade
+[ -e "$CONF.lenny-upgrade" ] || cp $CONF $CONF.lenny-upgrade
 sed -e 's|postgresql-8.1|postgresql-8.3|g
         s|nagios2|nagios3|g' -i $CONF
 
@@ -366,18 +383,28 @@ echo
 echo "#############"
 echo "# apt-utils #"
 echo "#############"
-echo -e "\n\n" | aptitude -y install apt-utils tasksel debian-archive-keyring dpkg locales
+aptitude -y install apt-utils tasksel debian-archive-keyring dpkg locales
+tweak_apt
 aptitude update
 echo
 
 echo "##############"
 echo "# postgresql #"
 echo "##############"
+# tweaking kdm
+CONF=/etc/init.d/kdm
+if [ -e "$CONF" ]; then
+ mv $CONF $CONF.lenny-upgrade
+ echo "#!/bin/sh" > $CONF
+ echo "exit 0" >> $CONF
+ chmod 755 $CONF
+fi
+aptitude -y install postgresql
 for i in postgresql postgresql-8.3 postgresql-client-8.3; do
- echo -e "\n\n" | aptitude -y install $i
  # check installed ok
- dpkg -s $i | grep -q ^"Status: install ok installed" || echo -e "\n\n" | aptitude -y reinstall $i
+ dpkg -s $i | grep -q ^"Status: install ok installed" || aptitude -y install $i
 done
+[ -e "$CONF.lenny-upgrade" ] && mv $CONF.lenny-upgrade $CONF
 /etc/init.d/postgresql-8.3 stop
 pg_dropcluster 8.3 main &> /dev/null
 pg_createcluster 8.3 main
@@ -434,32 +461,60 @@ for i in ldap moodle mrbs pykota; do
  echo
 done
 
+echo "###############"
+echo "# base-passwd #"
+echo "###############"
+tweak_apt
+aptitude -y install passwd
+aptitude -y install base-passwd
+echo
+
+if [ -n "$BITTORRENT" ]; then
+ echo "##############"
+ echo "# bittorrent #"
+ echo "##############"
+ if ! grep -q ^bittorrent: /etc/group; then
+  groupadd -r bittorrent
+ fi
+ if ! grep -q ^bittorrent: /etc/passwd; then
+  useradd -r -d /home/bittorrent -c "bittorrent user" -g bittorrent -s /bin/bash bittorrent
+ fi 
+ aptitude -y install bittorrent
+ chown bittorrent /var/log/bittorrent -R
+ chown bittorrent /var/lib/bittorrent -R
+ echo
+fi
+
 echo "################"
 echo "# dist-upgrade #"
 echo "################"
-echo -e "\n\n" | aptitude -y safe-upgrade
-echo -e "\n\n" | aptitude -y dist-upgrade
-echo -e "\n\n" | aptitude -y dist-upgrade
-echo -e "\n\n" | aptitude -y purge avahi-daemon
+tweak_apt
+aptitude -y safe-upgrade
+aptitude -y dist-upgrade
+aptitude -y dist-upgrade
+aptitude -y purge avahi-daemon
 echo
 
 echo "##############"
 echo "# sophomorix #"
 echo "##############"
+tweak_apt
 rm $INSTALLED
-echo -e "\n\n" | aptitude -y install $SOPHOPKGS || true
+aptitude -y install sophomorix2
 touch $INSTALLED
 echo
 
 echo "###############"
 echo "# common task #"
 echo "###############"
+tweak_apt
 linuxmuster-task --unattended --install=common
 echo
 
 echo "###############"
 echo "# server task #"
 echo "###############"
+tweak_apt
 linuxmuster-task --unattended --install=server
 echo
 
@@ -488,6 +543,7 @@ echo
 echo "################"
 echo "# imaging task #"
 echo "################"
+tweak_apt
 linuxmuster-task --unattended --install=imaging-$imaging
 echo
 
@@ -495,10 +551,11 @@ if [ -n "$FREERADIUS" ]; then
  echo "##########################"
  echo "# linuxmuster-freeradius #"
  echo "##########################"
+ tweak_apt
  aptitude -y install freeradius freeradius-ldap
  aptitude -y install linuxmuster-freeradius
  CONF=/etc/freeradius/clients.conf
- if [ -s "$CONF" -a -d "$FREEDYNTPLDIR" ]; then
+ if [ -s "$CONF" -a -d "$FREEDYNTPLDIR" -a ! -e "$CACHEDIR/.freeradius.upgrade50.done" ]; then
   echo "Aktualisiere freeradius ..."
   # fetch radiussecret
   found=false
@@ -526,6 +583,7 @@ if [ -n "$FREERADIUS" ]; then
    chmod 640 $targetcfg
    chown root:freerad $targetcfg
   done # targets
+  touch $CACHEDIR/.freeradius.upgrade50.done
  fi
  echo
 fi
@@ -542,7 +600,37 @@ if [ -n "$PYKOTA" ]; then
  echo "##################"
  echo "# linuxmuster-pk #"
  echo "##################"
+ tweak_apt
  aptitude -y install linuxmuster-pk
+ echo
+fi
+
+if [ -n "$PHPMYADMIN" ]; then
+ echo "##############"
+ echo "# phpmyadmin #"
+ echo "##############"
+ tweak_apt
+ aptitude -y install phpmyadmin
+ echo
+fi
+
+if [ -n "$PHPPGADMIN" ]; then
+ echo "##############"
+ echo "# phppgadmin #"
+ echo "##############"
+ tweak_apt
+ aptitude -y install phppgadmin
+ echo
+fi
+
+if [ -n "$OPENML" ]; then
+ echo "##########"
+ echo "# openML #"
+ echo "##########"
+ tweak_apt
+ echo "deb http://www.linuxmuster.net/openlml-unsupported/ openlml/" > /etc/apt/sources.list.d/openml.list
+ aptitude update
+ aptitude -y install linuxmuster-schulkonsole-templates-openlml
  echo
 fi
 
@@ -557,13 +645,16 @@ NAGUPGRADE=/usr/share/doc/nag2/examples/scripts/upgrades/2.1_to_2.2.sql
 TURBAUPGRADE=/usr/share/doc/turba2/examples/scripts/upgrades/2.1_to_2.2_add_sql_share_tables.sql
 for i in $HORDEUPGRADE $KRONOUPGRADE $MNEMOUPGRADE $NAGUPGRADE $TURBAUPGRADE; do
  t="$(echo $i | awk -F\/ '{ print $5 }')"
- if [ ! -s "$i" ]; then
-  echo " Fatal: $i not found! Cannot upgrade $t!"
+ if [ -e "$CACHEDIR/.${t}.upgrade50.done" ]; then
+  echo "$t wurde schon aktualisiert. Überspringe $t."
+  continue
+ elif [ ! -s "$i" ]; then
+  echo " Fehler: $i nicht gefunden! Überspringe $t!"
   continue
  else
-  echo " Upgrading $t ..."
+  echo " Aktualisiere $t ..."
  fi
- mysql horde < $i
+ mysql horde < $i && touch $CACHEDIR/.${t}.upgrade50.done
 done
 echo
 
@@ -574,6 +665,7 @@ echo "#############"
 rm -f /etc/apt/apt.conf.d/99upgrade
 # final stuff
 dpkg-reconfigure linuxmuster-base
+linuxmuster-nagios-setup
 echo
 
 echo "######################"
